Basic integration working, not sure about physics
