+++ package.json
@@ -12,8 +12,8 @@
     "lint": "prettier . --check && eslint .",
     "lint:fix": "prettier . --write && eslint . --fix",
     "test": "npm run lint:fix && npm run test:unit",
-    "test:dist": "cross-env TS_NODE_COMPILER_OPTIONS=\"{ \\\"module\\\": \\\"commonjs\\\" }\" mocha -r ts-node/register test/dist/**/*.ts",
-    "test:unit": "cross-env TS_NODE_COMPILER_OPTIONS=\"{ \\\"module\\\": \\\"commonjs\\\" }\" mocha -r ts-node/register test/unit/**/*.ts"
+    "test:dist": "cross-env TS_NODE_COMPILER_OPTIONS=\"{ \\\"module\\\": \\\"commonjs\\\" }\" mocha -r ts-node/register test/dist/**/*.spec.ts",
+    "test:unit": "cross-env TS_NODE_COMPILER_OPTIONS=\"{ \\\"module\\\": \\\"commonjs\\\" }\" mocha -r ts-node/register test/unit/**/*.spec.ts"
   },
   "repository": {
     "type": "git",
